backend/requirements.txt
fastapi
uvicorn[standard]
redis
google-generativeai
python-dotenv
pydantic
httpx
